+++ lib/aca_entities/policies/aptc_maximum.rb
@@ -6,8 +6,8 @@ module AcaEntities
     class AptcMaximum < Dry::Struct
       attribute :start_on,      Types::Date.optional.meta(omittable: true)
       attribute :end_on,        Types::Date.optional.meta(omittable: true)
-      attribute :max_aptc,      Types::Decimal.optional.meta(omittable: true)
-      attribute :aptc_percent,  Types::Decimal.optional.meta(omittable: true)
+      attribute :max_aptc,      Types::Float.optional.meta(omittable: true)
+      attribute :aptc_percent,  Types::Float.optional.meta(omittable: true)
     end
   end
 end
